--- app/src/main/java/kitchenpos/application/order/dto/OrderLineItemCommand.java
@@ -1,20 +1,0 @@
-package kitchenpos.application.order.dto;
-
-public class OrderLineItemCommand {
-
-    private Long menuId;
-    private long quantity;
-
-    public OrderLineItemCommand(Long menuId, long quantity) {
-        this.menuId = menuId;
-        this.quantity = quantity;
-    }
-
-    public Long menuId() {
-        return menuId;
-    }
-
-    public long quantity() {
-        return quantity;
-    }
-}
