--- app/src/main/java/kitchenpos/domain/order/OrderLineItem.java
@@ -1,46 +1,0 @@
-package kitchenpos.domain.order;
-
-import static javax.persistence.GenerationType.IDENTITY;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class OrderLineItem {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long seq;
-
-    private Long menuId;
-
-    @Column(nullable = false)
-    private long quantity;
-
-    protected OrderLineItem() {
-    }
-
-    public OrderLineItem(Long menuId, long quantity) {
-        this(null, menuId, quantity);
-    }
-
-    public OrderLineItem(Long seq, Long menuId, long quantity) {
-        this.seq = seq;
-        this.menuId = menuId;
-        this.quantity = quantity;
-    }
-
-    public Long seq() {
-        return seq;
-    }
-
-    public Long menuId() {
-        return menuId;
-    }
-
-    public long quantity() {
-        return quantity;
-    }
-}
