--- app/src/main/java/kitchenpos/domain/order/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.domain.order;
-
-public enum OrderStatus {
-    COOKING, MEAL, COMPLETION
-}
