--- app/src/main/java/kitchenpos/exception/order/OrderException.java
@@ -1,18 +1,0 @@
-package kitchenpos.exception.order;
-
-import kitchenpos.exception.BaseException;
-import kitchenpos.exception.BaseExceptionType;
-
-public class OrderException extends BaseException {
-
-    private final BaseExceptionType exceptionType;
-
-    public OrderException(OrderExceptionType exceptionType) {
-        this.exceptionType = exceptionType;
-    }
-
-    @Override
-    public BaseExceptionType exceptionType() {
-        return exceptionType;
-    }
-}
