+++ app/src/test/java/kitchenpos/application/IntegrationTest.java
@@ -5,9 +5,6 @@ import java.util.Arrays;
 import kitchenpos.application.menu.MenuGroupService;
 import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.menu.ProductService;
-import kitchenpos.application.order.OrderService;
-import kitchenpos.application.order.TableEmptyChangedEventListener;
-import kitchenpos.application.order.TableUngroupedEventListener;
 import kitchenpos.application.table.OrderedEventListener;
 import kitchenpos.application.table.TableGroupService;
 import kitchenpos.application.table.TableService;
@@ -20,15 +17,18 @@ import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.menu.Price;
 import kitchenpos.domain.menu.Product;
 import kitchenpos.domain.menu.ProductRepository;
-import kitchenpos.domain.order.Order;
-import kitchenpos.domain.order.OrderLineItem;
-import kitchenpos.domain.order.OrderLineItems;
-import kitchenpos.domain.order.OrderRepository;
-import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
 import kitchenpos.domain.table.TableGroup;
 import kitchenpos.domain.table.TableGroupRepository;
+import kitchenpos.order.application.OrderService;
+import kitchenpos.order.application.TableEmptyChangedEventListener;
+import kitchenpos.order.application.TableUngroupedEventListener;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItems;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.springframework.beans.factory.annotation.Autowired;
