+++ app/src/test/java/kitchenpos/application/table/TableGroupServiceTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.application.table;
 
-import static kitchenpos.domain.order.OrderStatus.COOKING;
 import static kitchenpos.exception.table.OrderTableExceptionType.ORDER_TABLE_NOT_FOUND;
-import static kitchenpos.exception.table.TableGroupExceptionType.CAN_NOT_UNGROUP_COOKING_OR_MEAL;
 import static kitchenpos.exception.table.TableGroupExceptionType.ORDER_TABLES_CAN_NOT_LESS_THAN_TWO;
-import static kitchenpos.exception.table.TableGroupExceptionType.ORDER_TABLE_MUST_EMPTY;
 import static kitchenpos.exception.table.TableGroupExceptionType.ORDER_TABLE_CAN_NOT_HAVE_TABLE_GROUP;
+import static kitchenpos.exception.table.TableGroupExceptionType.ORDER_TABLE_MUST_EMPTY;
+import static kitchenpos.order.domain.OrderStatus.COOKING;
+import static kitchenpos.order.exception.OrderExceptionType.ORDER_STATUS_IS_COOKING_OR_MEAL;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -141,7 +141,7 @@ class TableGroupServiceTest extends IntegrationTest {
         ).exceptionType();
 
         // then
-        assertThat(exceptionType).isEqualTo(CAN_NOT_UNGROUP_COOKING_OR_MEAL);
+        assertThat(exceptionType).isEqualTo(ORDER_STATUS_IS_COOKING_OR_MEAL);
     }
 
     @Test
