+++ app/src/test/java/kitchenpos/ui/ControllerTest.java
@@ -8,7 +8,6 @@ import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.menu.ProductService;
 import kitchenpos.application.menu.dto.CreateMenuResponse;
 import kitchenpos.application.menu.dto.SearchMenuResponse;
-import kitchenpos.application.order.OrderService;
 import kitchenpos.application.table.TableGroupService;
 import kitchenpos.application.table.TableService;
 import kitchenpos.domain.menu.Menu;
@@ -17,12 +16,13 @@ import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.menu.MenuProducts;
 import kitchenpos.domain.menu.Price;
 import kitchenpos.domain.menu.Product;
-import kitchenpos.domain.order.Order;
-import kitchenpos.domain.order.OrderLineItem;
-import kitchenpos.domain.order.OrderLineItems;
-import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.TableGroup;
+import kitchenpos.order.application.OrderService;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItems;
+import kitchenpos.order.domain.OrderStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
