+++ order/build.gradle
@@ -1,4 +1,3 @@
 dependencies {
     implementation project(':common')
-    implementation project(':order')
 }
