+++ order/src/main/java/kitchenpos/order/application/dto/ChangeOrderStatusCommand.java
@@ -1,0 +1,22 @@
+package kitchenpos.order.application.dto;
+
+import kitchenpos.order.domain.OrderStatus;
+
+public class ChangeOrderStatusCommand {
+
+    private final Long orderId;
+    private final OrderStatus orderStatus;
+
+    public ChangeOrderStatusCommand(Long orderId, OrderStatus orderStatus) {
+        this.orderId = orderId;
+        this.orderStatus = orderStatus;
+    }
+
+    public Long orderId() {
+        return orderId;
+    }
+
+    public OrderStatus orderStatus() {
+        return orderStatus;
+    }
+}
