+++ order/src/main/java/kitchenpos/order/application/dto/OrderLineItemResponse.java
@@ -1,0 +1,40 @@
+package kitchenpos.order.application.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import kitchenpos.order.domain.OrderLineItem;
+
+public class OrderLineItemResponse {
+
+    @JsonProperty("seq")
+    private Long seq;
+    @JsonProperty("menuId")
+    private Long menuId;
+    @JsonProperty("quantity")
+    private long quantity;
+
+    private OrderLineItemResponse(Long seq, Long menuId, long quantity) {
+        this.seq = seq;
+        this.menuId = menuId;
+        this.quantity = quantity;
+    }
+
+    public static OrderLineItemResponse from(OrderLineItem orderLineItem) {
+        return new OrderLineItemResponse(
+                orderLineItem.seq(),
+                orderLineItem.menuId(),
+                orderLineItem.quantity()
+        );
+    }
+
+    public Long seq() {
+        return seq;
+    }
+
+    public Long menuId() {
+        return menuId;
+    }
+
+    public long quantity() {
+        return quantity;
+    }
+}
