+++ order/src/main/java/kitchenpos/order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.domain;
+
+public enum OrderStatus {
+    COOKING, MEAL, COMPLETION
+}
